购物车，地址簿，订单提交

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -50,8 +50,8 @@ public class CategoryServiceImpl implements CategoryService {
         //设置创建时间、修改时间、创建人、修改人
         category.setCreateTime(LocalDateTime.now());
         category.setUpdateTime(LocalDateTime.now());
-        category.setCreateUser((Long) BaseContext.getCurrentId());
-        category.setUpdateUser((Long) BaseContext.getCurrentId());
+        category.setCreateUser(BaseContext.getCurrentId());
+        category.setUpdateUser(BaseContext.getCurrentId());
 
         categoryMapper.insert(category);
     }
@@ -101,7 +101,7 @@ public class CategoryServiceImpl implements CategoryService {
 
         //设置修改时间、修改人
         category.setUpdateTime(LocalDateTime.now());
-        category.setUpdateUser((Long) BaseContext.getCurrentId());
+        category.setUpdateUser(BaseContext.getCurrentId());
 
         categoryMapper.update(category);
     }
@@ -116,7 +116,7 @@ public class CategoryServiceImpl implements CategoryService {
                 .id(id)
                 .status(status)
                 .updateTime(LocalDateTime.now())
-                .updateUser((Long) BaseContext.getCurrentId())
+                .updateUser(BaseContext.getCurrentId())
                 .build();
         categoryMapper.update(category);
     }
